chore: StyledBottomTabBar 컴포넌트 위치 이동 	renamed:    src/navigation/components/StyledBottomTabBar.tsx -> src/navigation/components/styledBottomTabBar/StyledBottomTabBar.tsx 	renamed:    src/navigation/components/StyledBottomTabBarButton.tsx -> src/navigation/components/styledBottomTabBar/StyledBottomTabBarButton.tsx 	renamed:    src/navigation/components/StyledBottomTabBarIcon.tsx -> src/navigation/components/styledBottomTabBar/StyledBottomTabBarIcon.tsx 	renamed:    src/navigation/components/StyledBottomTabBarLabel.tsx -> src/navigation/components/styledBottomTabBar/StyledBottomTabBarLabel.tsx 	renamed:    src/navigation/components/StyledNavigationContainer.tsx -> src/navigation/components/styledBottomTabBar/StyledNavigationContainer.tsx 	renamed:    src/navigation/components/__test__/StyledBottomTabBar.test.tsx -> src/navigation/components/styledBottomTabBar/__test__/StyledBottomTabBar.test.tsx 	renamed:    src/navigation/components/__test__/StyledBottomTabBarButton.test.tsx -> src/navigation/components/styledBottomTabBar/__test__/StyledBottomTabBarButton.test.tsx 	renamed:    src/navigation/components/__test__/StyledBottomTabBarLabel.test.tsx -> src/navigation/components/styledBottomTabBar/__test__/StyledBottomTabBarLabel.test.tsx 	renamed:    src/navigation/components/__test__/StyledNavigationContainer.test.tsx -> src/navigation/components/styledBottomTabBar/__test__/StyledNavigationContainer.test.tsx
