+++ src/navigation/components/styledBottomTabBar/StyledBottomTabBar.tsx
