+++ src/navigation/components/styledBottomTabBar/StyledBottomTabBarIcon.tsx
