+++ src/navigation/components/styledBottomTabBar/StyledBottomTabBarLabel.tsx
