+++ src/navigation/components/styledBottomTabBar/StyledNavigationContainer.tsx
