+++ src/navigation/components/styledBottomTabBar/__test__/StyledBottomTabBarLabel.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render, screen } from '@testing-library/react-native';
 
-import StyledBottomTabBarLabel from '@navigation/components/StyledBottomTabBarLabel';
+import StyledBottomTabBarLabel from '@navigation/components/styledBottomTabBar/StyledBottomTabBarLabel';
 import { settingStore } from '@store/settingStore/useSettingStore';
 
 import type { LocalizedText } from '@libs/utils/localize/localize.type';
